Fix 44 warnings for alibgp (151-194) and enable strict compiling for debian (195)
- 151 pathd: include path_debug.h in path_main.c for pathd_debug_init
- 152 pathd: change param dnode type to const for cli_show_srte_policy_binding_v6_sid
- 153 pathd: change param dnode type to const for cli_show_srte_policy_sbfd
- 154 pathd: remove unused variables in pathd_srte_segment_list_segment_v6_sid_value_modify
- 155 pathd: remove unused variables in pathd_srte_policy_candidate_path_segment_list_name_modify
- 156 pathd: change 'G' as variable in sidlist_Db_SetEntry
- 157 pathd: change 'G' to variable in sidlist_Db_DelEntry
- 158 pathd: add prototype for path_zebra_encode_srv6_policy
- 159 pathd: remove const for param msg for g_sidlist_appdb_redis.redis_PublishMsg
- 160 pathd: remove unused variable in srte_policy_best_candidate
- 161 pathd: add break for switch case in cpath_status_down_handle
- 162 pathd: use strlcpy in srte_candidate_bfd_group_find
- 163 pathd: use strlcpy in srte_candidate_bfd_group_add_with_status
- 164 pathd: fix non-return issue for srte_candidate_bfd_group_add
- 165 pathd: use strlcpy in srte_candidate_bfd_group_add
- 166 staticd: remove unmatched specifier for vty_out in static_route_leak
- 167 staticd: remove unused variable in ip_route_v6gate_vrf_magic
- 168 staticd: fix param issue in ipv6_route_evpn_vrf_magic
- 169 staticd: add const for rmac in nexthop_cli_show
- 170 staticd: add const for param vrfname of static_route_show
- 171 bfdd: change return type to const for bfd_mode_type_to_string
- 172 bfdd: fix vty_out specifier issue in _display_peer
- 173 bfdd: change buf type to const in _display_peer_brief
- 174 bfdd: fix specifier type issue for vty_out in _display_peer_brief
- 175 bfdd: change param vrfname and bfdname type to const for _display_bfd_by_bfdname
- 176 bfdd: change param vrfname and bfdname type to const for _display_bfd_counters_by_bfdname
- 177 bfdd: change type of param vrfname and bfdname to const for _clear_bfd_counters_by_bfdname
- 178 staticd: add break for switch case in nexthop_cli_show
- 179 bfdd: fix specifier issue for snprintf in bfd_peer_enter_magic
- 180 bfdd: change determin_ip_version to static to static
- 181 staticd: change _bfd_cli_bfd_mode_type_to_string return type to const
- 182 staticd: fix specifier issue for vty_out in nexthop_cli_show
- 183 staticd: add case for switch in static_nexthop_cli_cmp
- 184 bfdd: fix pointer issue in _display_peers_brief
- 185 zebra: fix var type issue in zsend_neighbor_notify
- 186 bfdd: fix var type issue and remove unused var in bfd_session_create
- 187 bfdd: change seglist_name_len type as uint16_t
- 188 bfdd: fix packed member unaligned issue in _ptm_sbfd_send
- 189 bfdd: fix packed member unaligned issue in _ptm_sbfd_echo_send
- 190 bfdd: extend var field in bfd_Db_GetSessStatus
- 191 pathd: convert params to non-const in srte_sbfd_session_compare
- 192 pathd: add prototype for _sbfd_candidate_seglist_disable
- 193 pathd: remove unused function sbfd_pathd_candidate_removed_handler
- 194 zebra: include zebra_db.h in zebra_srv6_vty.c for zebra_Db_Set_SRV6_LOCAL_ENDX_SID
- 195 debian: enable Werror and start strict compiling for debian

+++ bfdd/bfdctl.h
@@ -122,7 +122,7 @@ struct bfd_peer_cfg {
     /* sbfd */
 	bool bpc_sbfd;
 	uint8_t seg_num;
-	uint8_t seglist_name_len;
+	uint16_t seglist_name_len;
 	char seglist_name[MAXNAMELEN + 1];
 	uint32_t srte_color;
 	struct in6_addr srte_endpoint;
